Move permissions into its own reducer
